Add P2 detection head, scene augmentation, DIoU-NMS with ablation evaluation
- yolov13n-apple-p2.yaml: 4-head model (P2/P3/P4/P5) with stride-4 branch from backbone layer 2
- apple_orchard_nano.yaml: tuned training config with copy_paste=0.15, mixup=0.05, hsv_v=0.4, DIoU-NMS
- augment.py: scene-targeted augmentation modifications for orchard lighting variation
- eval_*.py, count_all_variants.py: ablation evaluation scripts (detection, counting, NMS comparison, per-size recall)
- draw_arch.py: architecture diagram generator showing improvements
- thesis_output/: result figures for all ablation variants A0-A5
- README.md: project documentation and replication guide

+++ yolov13-main/ultralytics/data/augment.py
@@ -943,7 +943,7 @@ class MixUp(BaseMixTransform):
         """
         r = np.random.beta(32.0, 32.0)  # mixup ratio, alpha=beta=32.0
         labels2 = labels["mix_labels"][0]
-        labels["img"] = (labels["img"] * r + labels2["img"] * (1 - r)).astype(np.uint8)
+        labels["img"] = (labels["img"].astype(np.float32) * r + labels2["img"].astype(np.float32) * (1 - r)).astype(np.uint8)
         labels["instances"] = Instances.concatenate([labels["instances"], labels2["instances"]], axis=0)
         labels["cls"] = np.concatenate([labels["cls"], labels2["cls"]], 0)
         return labels
